refactored nav into header types
