+++ pizza-time/src/components/user-dashboard/user-dashboard.js
@@ -1,14 +1,13 @@
 import React from 'react';
 
-import Nav from '../nav/nav.js';
+import Nav from '../home-header/home-header.js';
 import Footer from '../footer/footer.js';
 
 const UserDashboard = () => {
     return(
         <div>
             <Nav />
-            User DashBoard Works!
-            <Footer />
+            {/* <Footer /> */}
         </div>
     );
 }
