+++ pizza-time/src/styles/navStyles.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Wrap = styled.div`
+  width: 100%;
+  border: 1px solid red;
+`;
+
+  export const Inner = styled.div`
+    width: 100%;
+    padding: 15px 0;
+  `;
